Refactored and move new functionality to core dir
