+++ src/core/handler.container.ts
@@ -7,9 +7,8 @@ export class HandlerContainer {
 	private container: Map<string, { provider: Provider; methodName: string }> = new Map();
 
 	public register(actionOrIntent: string, provider: Provider, methodName: string): void {
-    const registeredHandler = this.container.get(actionOrIntent);
-    
-    if (registeredHandler) {
+
+    if (this.container.has(actionOrIntent)) {
       throw new Error(`Cannot have duplicate handlers for intent [${actionOrIntent}]`);
     }
 
@@ -20,11 +19,12 @@ export class HandlerContainer {
 		actionOrIntent: string,
 		dialogFlowResponse: DialogFlowResponse,
 	): Promise<DialogFlowFulfillmentResponse> {
-		const registeredHandler = this.container.get(actionOrIntent);
 
-		if (!registeredHandler) {
+		if (!this.container.has(actionOrIntent)) {
 			throw new Error(`Unknown handler for [${actionOrIntent}].`);
 		}
+
+		const registeredHandler = this.container.get(actionOrIntent);
 
 		const { provider, methodName } = registeredHandler;
